# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Policy Platform API",
    "description": "Enterprise policy ingestion and immutable versioning service. Provides ingestion batches, blob registration, queue-driven extraction, and audit-friendly traceability.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/policies": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List policies",
        "description": "Lists policies for a tenant.",
        "operationId": "list_policies_v1_policies_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyResponse"
                  },
                  "title": "Response List Policies V1 Policies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policies/{policy_id}/versions": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List policy versions",
        "description": "Lists immutable policy versions for a policy (tenant-scoped).",
        "operationId": "list_policy_versions_v1_policies__policy_id__versions_get",
        "parameters": [
          {
            "name": "policy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicyVersionResponse"
                  },
                  "title": "Response List Policy Versions V1 Policies  Policy Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/policy-versions/{policy_version_id}/sections": {
      "get": {
        "tags": [
          "Policies"
        ],
        "summary": "List extracted sections",
        "description": "Lists extracted sections for a policy version (tenant-scoped).",
        "operationId": "list_policy_version_sections_v1_policy_versions__policy_version_id__sections_get",
        "parameters": [
          {
            "name": "policy_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Policy Version Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Tenant UUID",
              "title": "Tenant Id"
            },
            "description": "Tenant UUID"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max number of sections to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max number of sections to return"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PolicySectionResponse"
                  },
                  "title": "Response List Policy Version Sections V1 Policy Versions  Policy Version Id  Sections Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PolicyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "current_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "updated_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By User Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "external_id",
          "name",
          "status",
          "jurisdiction",
          "category",
          "current_version_id",
          "created_at",
          "created_by_user_id",
          "updated_at",
          "updated_by_user_id"
        ],
        "title": "PolicyResponse"
      },
      "PolicySectionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "policy_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Version Id"
          },
          "section_index": {
            "type": "integer",
            "title": "Section Index"
          },
          "section_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Path"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "start_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Offset"
          },
          "end_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Offset"
          },
          "content_sha256": {
            "type": "string",
            "title": "Content Sha256"
          },
          "rag_document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rag Document Id"
          },
          "rag_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rag Node Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "policy_version_id",
          "section_index",
          "section_path",
          "title",
          "text",
          "start_offset",
          "end_offset",
          "content_sha256",
          "rag_document_id",
          "rag_node_id",
          "created_at"
        ],
        "title": "PolicySectionResponse"
      },
      "PolicyVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "policy_id": {
            "type": "string",
            "format": "uuid",
            "title": "Policy Id"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "version_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Label"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "effective_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Date"
          },
          "blob_container": {
            "type": "string",
            "title": "Blob Container"
          },
          "blob_name": {
            "type": "string",
            "title": "Blob Name"
          },
          "raw_blob_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Blob Uri"
          },
          "blob_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Version Id"
          },
          "blob_etag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Etag"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "content_length": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Length"
          },
          "extracted_blob_container": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Blob Container"
          },
          "extracted_blob_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Blob Name"
          },
          "extracted_blob_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Blob Uri"
          },
          "content_sha256": {
            "type": "string",
            "title": "Content Sha256"
          },
          "metadata_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata Json"
          },
          "metadata_sha256": {
            "type": "string",
            "title": "Metadata Sha256"
          },
          "ingest_batch_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingest Batch Id"
          },
          "parse_status": {
            "type": "string",
            "title": "Parse Status"
          },
          "parse_status_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Parse Status Updated At"
          },
          "parse_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Error Code"
          },
          "parse_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Error Message"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current",
            "default": false
          },
          "supersedes_policy_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Policy Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by_user_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By User Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "policy_id",
          "version_number",
          "title",
          "effective_date",
          "blob_container",
          "blob_name",
          "blob_version_id",
          "blob_etag",
          "content_type",
          "content_length",
          "content_sha256",
          "metadata_json",
          "metadata_sha256",
          "ingest_batch_id",
          "parse_status",
          "parse_status_updated_at",
          "parse_error_code",
          "parse_error_message",
          "created_at",
          "created_by_user_id",
          "correlation_id"
        ],
        "title": "PolicyVersionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}